templates/python-api-gateway/requirements.txt
aws-cdk.core

aws-cdk.aws_apigateway